Added my brightness powershell script
A simple bit of code that will adjust the brightness of screen when run.

To run it forever, create a new task in the task scheduler and do the following:
Tab 1: Select "Run when user is logged on" and "Hidden"
Tab 2: Create a trigger that starts running when the user logs in. You can adjust this as needed
Tab 3:Create an action, and enter in as arguments: "-NoProfile -WindowStyle Hidden -ExecutionPolicy Bypass -file "<path to file>\brightness.ps1". Technically you don't need -WindowStyle Hidden, as you select that in the first tab, but I have it here anyways.
Tab 4/Tab 5: Do whatever you want

And it should run perfectly! (nearly so)